backend/requirements/base.txt
boto3==1.14.9
Django==3.0.7
django-constance==2.7.0
django-environ==0.4.5
django-extensions==2.2.9
django-picklefield==3.0.1
django-redis==4.12.1
django-storages==1.9.1
djangorestframework==3.11.0
gunicorn==20.0.4
ipython==7.15.0
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
requests==2.24.0
social-auth-app-django==4.0.0
tqdm==4.46.1
uvicorn==0.11.5
